Player: Improve resting
Instead of making GoDot process extra fast which creates strange
behaviour if a monster has appeared but not attacked yet, use
an accelerator that is applied to time (like walking).

The accelerator is only applied as long as there are no hostiles.
If someone is hostile then time slows down to the normal speed
and you can watch the hostile come in to attack you.

If a hostile attacks the player, the player should wake up even
if they aren't hit.  This gives the attacker a free surprise attack
but doesn't give them a free hit.

Don't allow resting when there are hostiles about.