Navigation and spawing fixes
I hit a problem when I tried to have the KoboldWarrior spawning
kobolds highlighting that there are big problems with how I was
handling object placement and navigation.

Redo all of this to try to keep the navigation logic in the physics
frame.

As a consequence of doing this, I ended up redoing the persistent
state of spawned items.

I also changed the logic so friendly actors can move as well
and doing so I had to make sure that things stop navigating when
their global_position changes.  They can reacquire their destination
at that point if it still makes sense.