GameEngine: Add method to spawn a number of the same scene
This is intended for spawning monsters as it will spawn them
close to the player.

Take care to disable physics while doing the spawning and
also keep track and exclude the places used by earlier spawned
scenes.

This fixes a problem where 2 monsters are spawned at the same
location causing the physics processing to go insane and
blast them wildly off the screen.