Ephemeral group: don't go crazy with the delete-whizz!
Previously every single thing in the scene needed to be in the
game save file or it would be removed.

This includes very useful things like doors, message areas, etc.

Instead, mark nodes that can appear and disappear from the scene
with the group "Ephemeral" and only those nodes should be deleted
when they are contained in the game save file.