Things: Only inventory things should be finable, add a larger shape too
Move  the findable logic up from Thing to InventoryThing and add a
collision shape for finding it which is (by default) larger than the
normal interaction shape.

This makes it easier to find and to place things that you want to
be found while still leaving the ability to make things really
hard to find.