Actor/Player: Unify the attack code
Add hooks for the player (or other monster types) to be selective
about who they attack.

Get rid of the special player logic to handle attacks.

Add an Attack (or Attacks) as children of InventoryItems that
have an attack property.

Add a range for attacks so that wolves can attack from further
away (for example) and different weapons can have different ranges.

Change Thing to take a small amount to actually do the use of an item
and the same time between uses.

This is all a precursor for potentially implementing ranged wweapons.