random encounters: Fix broken spawning from Godot 4.0 conversion
All spawners were failing to spawn and being left at their default (0,0)
position.

Fix this by checking to see if we could place the monster.  If we couldn't
place it, then delete it.

Fix the placing error from godot 4.0 conversion which was to reverse the
direction of the ray checking if we could see the player from the position
we were wantin to place the monters.